Support --json option for transform.
Which now combines the static module record meta data with the functor
program as a JSON document.

The aim is to call this from the endo compartment-mapper to determine if
this implementation is correct.

Note that we needed to jump through some hoops to convert the initial
static module record into an owned version as it was borrowing from both
the parser/analyzer structs and from the source module AST.