Hack for member expressions that match locals.
This is a quick fix for the issue where member expressions were being
included when the first part of the expression path should have matched
a local declared in the same or higher lexical scope but instead was not
matching and being treated as a global incorrectly.

Fix a test spec by re-ordering the expected output which needs a proper
fix to maintain the correct order.