Improve member expression handling.
Cater for the expression:

[].slice.call(arguments)

Which yields the array literal as the deepest member expression object
which is inconvenient for our iteration order so we keep track of member
expressions and discard the entire expression if the deepest member
expression object is a literal value.

Remove `this` normalization test spec as we will ignore globals accessed
via `this` for the moment as it would require cross-module analysis.