Fix nested member expression ident handling.
Using visit_ident() is too greedy because it also includes object keys
so we need to manually walk the parts of the AST that interest us.

Note that lots more work could be done on the logic for visiting nested
member expressions to make the iteration more correct.

Currently, it only covers the existing test cases.